Online shopping intension
Analyzing the online purchasing intention of a customer has been depicted in the form of a paper, where we have used acquired datasets to gather data consisting of numerical and categorical variables respectively. On completion of Data collection, this model includes KNN model to predict the probability of online shoppers to purchase a product.